requirements.txt
asgiref==3.2.10
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-crispy-forms==1.9.2
django-import-export==2.4.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.4.0
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
